requirements_test.txt
pytest
pytest-homeassistant-custom-component
pyserial
psutil-home-assistant
fnv_hash_fast
pytz
tzdata
janus
Pillow
ruff
mypy
bandit
